GH-193: Always Run close() on the Event Thread
Resolves https://github.com/reactor/reactor-kafka/issues/193

Previously, `ConsumerEventLoop.stop()` ran the close event on the
current thread, if that thread is a receiver thread. However, the
current thread could be for a different receiver.

Always schedule the `CloseEvent` on the loop's scheduler, and wake
the consumer since the thread might be suspended in `poll()`.

Also fixes a race condition in another test.