[build] Check license headers with spotless
This commit introduces Spotless plugin to mainly check license headers,
in both java files and gradle files.

The gradle configuration is trickier, as a "top-of-file" marker must be
provided, but Gradle build files are very flexible.
The regular expression attempts detecting 3 patterns:
 - `import` statements
 - `simpleBlockStart {` (typically: `buildScript`)
 - `parameterizedBlock(start) {` (typically: `configure(rootProject)`)
It also accepts a space before the `{` character.

This commit also adds the license template under `codequality` folder,
and moves the checkstyle configuration under that same folder.

Finally, it performs the `spotless` checks as a preliminary step to
the ci workflow, ensuring a PR check will fail fast if there are any
Spotless violations.