KafkaReceiverTest.transactionalOffsetCommit Fixes
See https://github.com/reactor/reactor-kafka/issues/274

This test had several problems:

The second part of the test was receiving from the original topic instead
of the destination topic so it always succeeeded in finding all the records.

When this was changed to consume from the proper topic, it always failed.

This was because the `.take(count)` in `receiveAndSendTransactions` caused
the flux to be canceled before the final commit took place.

Added a callback hook to the `TransactionManager` so we can test that the commit
is complete before terminating the flux.

It is not clear whether these fixes will resolve the original problem so I
have left diagnostics for future failure analysis.

Also capture test results in the publish action.