Fix Pause During Back Pressure
If a user pauses partitions, then back pressure is applied, pausing all partitions,
the user paused partitions are not resumed when the back pressure is relieved.

However, if back pressure is in place when the user pauses partitions, they are
resumed when the back pressure is relieved.

Instead of determining user paused partitions dynamically, maintain a collection
thereof when the user invokes pause or resume on the consumer.