GH-314: Fix Cooperative Incremental Rebalance
Resolves https://github.com/reactor/reactor-kafka/issues/314

During cooperative rebalancing, e.g. with `CooperativeStickyAssignor`
commits can result in `RebalanceInProgressException`; this exception
represents a retryable condition, even though that exception does not
extend the existing `RetriableCommitFailedException` or its super interface
`RetriableException` so the event loop does not retry these exceptions.

Add this exception to the retryable exception list. Also, do not pause
the consumer during retries because it can continue to consume from partitions
that were not revoked by the cooperative rebalance.

Add unit test; integration testing was performed with the code provided
in the original issue #314.