GH-389: Fix MockReceiverTest#consumerMethods failure with reactor-core 3.5
Fixes: #389

Due to the change between `3.4.x` and `3.5.x` of `reactor-core`, the test would fail. 
The reason is that since `3.5` the prefetching strategy is less eager and the consumption would be paused due to lack of downstream requests. 
The assertion was to strict as in such scenarios not only the intended partition is paused but all of them. 
This relaxation makes the test pass against future reactor-core versions.